backend/requirements.txt
Django==4.2.13
APScheduler==3.10.4
pytz==2024.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
Pillow==10.3.0
gunicorn==21.2.0
openai==1.51.0
httpx>=0.27.0,<0.28.0
requests>=2.31.0
